Correcting errors. Moving _is_parametric to cunqa.circuit to avoid circular imports...
